requirements.txt
fastapi
uvicorn
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
beautifulsoup4
SQLAlchemy
pydantic[email]
bcrypt
python-jose
twilio
python-multipart
firebase-admin
APScheduler
pytest
pytest-cov
httpx
celery  # Optional
flower  # Optional
rabbitmq  # Optional
redis  # Optional
